requirements.txt
schematics==2.0.1
pandas==0.22.0
pyinterval==1.2.0
scipy==1.0.0
tqdm==4.19.5
sortedcontainers==1.5.9